src/test/javascript/spec/app/entities/agence-voyage/agence-voyage.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { AgenceVoyageService } from 'app/entities/agence-voyage/agence-voyage.service';
import { IAgenceVoyage, AgenceVoyage } from 'app/shared/model/agence-voyage.model';

describe('Service Tests', () => {
  describe('AgenceVoyage Service', () => {
    let injector: TestBed;
    let service: AgenceVoyageService;
    let httpMock: HttpTestingController;
    let elemDefault: IAgenceVoyage;
    let expectedResult: IAgenceVoyage | IAgenceVoyage[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(AgenceVoyageService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new AgenceVoyage(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a AgenceVoyage', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new AgenceVoyage()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a AgenceVoyage', () => {
        const returnedFromService = Object.assign(
          {
            nom: 'BBBBBB',
            lieu: 'BBBBBB',
            ville: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of AgenceVoyage', () => {
        const returnedFromService = Object.assign(
          {
            nom: 'BBBBBB',
            lieu: 'BBBBBB',
            ville: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a AgenceVoyage', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
